Extract funktor_options into a module we can inject into ActiveJob::Base
